# Conflicts:
#	README.md
#	src/main/java/pl/pabilo8/immersiveintelligence/api/crafting/ElectrolyzerRecipe.java
#	src/main/java/pl/pabilo8/immersiveintelligence/client/gui/overlay/gun/GuiOverlayGunBase.java
#	src/main/java/pl/pabilo8/immersiveintelligence/client/util/amt/AMTQuads.java
#	src/main/java/pl/pabilo8/immersiveintelligence/common/IICreativeTab.java
#	src/main/java/pl/pabilo8/immersiveintelligence/common/compat/TerrafirmaHelper.java
#	src/main/java/pl/pabilo8/immersiveintelligence/common/crafting/IIRecipes.java
#	src/main/java/pl/pabilo8/immersiveintelligence/common/item/tools/ItemIITrenchShovel.java
#	src/main/java/pl/pabilo8/immersiveintelligence/common/weaponsystem/IIWeaponUpgrade.java
#	src/main/java/pl/pabilo8/immersiveintelligence/common/weaponsystem/weapons/IIWeaponMachinegun.java

+++ gradle/tools/opti.sh
@@ -12,8 +12,14 @@ processAudio () {
     touch $1.tmp
     $LOC_FFMPEG -y -i $1 -f ogg -c:a libvorbis -b:a 120k $1.tmp &
     wait
-    rm $1
-    mv $1.tmp $1
+    if [ -s $1.tmp ]
+    then
+      rm $1
+      mv $1.tmp $1
+    else
+      echo "Cannot convert file, output is empty."
+      return;
+    fi
 }
 
 while getopts "ia" arg; do
